Removing Start - Stop Observer by Moving them into the getObserverMessage.
The reason is because of when observing (Scanning) on the TI chip, the TI chip could take as long as 100mSecs if not more to scan. During the scan process, the TI chip becomes busy with scanning and could ignore many messages over UART, therefore creating this block function so that we begin scanning when entering the function and stopping observing when exiting.